requirements.txt
 pandas
 openpyxl
 streamlit
 google-generativeai
 python-dotenv